# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Innoscripta News Aggregator API Docs'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://innoscripta-news-aggregator.test'
paths:
...
  /api/register:
    post:
      summary: Registration
      operationId: registration
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Insert name'
                  example: 'John Doe'
                  nullable: false
                email:
                  type: string
                  description: 'Insert email'
                  example: test@example.com
                  nullable: false
                password:
                  type: string
                  description: 'Insert password'
                  example: password
                  nullable: false
                password_confirmation:
                  type: string
                  description: 'Insert password confirmation'
                  example: password
                  nullable: false
              required:
                - name
                - email
                - password
                - password_confirmation
      security: []
...
  /api/password/reset:
    post:
      summary: 'Reset Password'
      operationId: resetPassword
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Insert email'
                  example: test@example.com
                  nullable: false
                password:
                  type: string
                  description: 'Insert password'
                  example: password
                  nullable: false
                password_confirmation:
                  type: string
                  description: 'Insert password confirmation'
                  example: password
                  nullable: false
              required:
                - email
                - password
                - password_confirmation
      security: []
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []